avatar-requirements.txt
flask==2.0.1
flask-cors==3.0.10
Pillow==9.5.0
opencv-python==4.7.0.72
numpy==1.24.3
rembg==2.0.50
werkzeug==2.0.1
